Update README.md
Show Box APK is a popular application for Android devices, designed to provide users with access to a wide range of movies and TV shows for streaming and downloading. It is known for its extensive library, user-friendly interface, and the convenience it offers in accessing entertainment content without the need for a subscription or payment.